+++ R/tqDist.r
@@ -1,4 +1,3 @@
-
 #' Status of quartets
 #' 
 #' Determines the number of quartets that are consistent within pairs of trees.
@@ -6,7 +5,7 @@
 #' Given a list of trees, returns the number of quartet statements present in the
 #' reference tree (the first entry in `trees`, if `cf` is not specified)
 #' that are also present in each other tree.  A random pair of fully resolved 
-#' trees is expected to share \code{choose(n_tip, 4) / 3} quartets.
+#' trees is expected to share `choose(n_tip, 4) / 3` quartets.
 #' 
 #' 
 #' If trees do not bear the same number of tips, `SharedQuartetStatus()` will 
@@ -23,6 +22,12 @@
 #' 
 #' @template treesParam
 #' @template treesCfParam
+#' @param nTip Integer specifying number of tips that could have occurred
+#' in `trees`.  Useful if comparing trees from different data sources that
+#' contain non-overlapping tips.
+#' If `NULL`, the default, then trees are assumed to contain the same tips.
+#' If `TRUE`, then a vector is generated automatically by counting all unique
+#' tip labels found in `trees` or `cf`.
 #' 
 #' @templateVar intro `QuartetStatus()` returns a two dimensional array. Rows correspond to the input trees; the first row will report a perfect match if the first tree is specified as the comparison tree (or if `cf` is not specified).  Columns list the status of each quartet:
 #' @template returnEstabrook
@@ -41,13 +46,20 @@
 #' 
 #' # Calculate Estabrook et al's similarity measures:
 #' SimilarityMetrics(sq_status)
+#' 
+#' # Compare trees that include a subset of the taxa 1..10
+#' library('TreeTools', quietly = TRUE, warn.conflict = FALSE)
+#' QuartetStatus(BalancedTree(1:5), BalancedTree(3:8), nTip = 10)
+#' 
+#' # If all taxa studied occur in `trees` or `cf`, set `nTip = TRUE`
+#' QuartetStatus(BalancedTree(1:5), BalancedTree(3:10), nTip = TRUE)
 #'  
 #' @family element-by-element comparisons
 #' @seealso 
-#' - [`SplitStatus()`]: Uses splits (groups/clades defined by
-#'  nodes or edges of the tree) instead of quartets as the unit of comparison.
+#' - Use splits (groups/clades defined by nodes or edges of the tree) instead
+#'   of quartets as the unit of comparison: [`SplitStatus()`].
 #'  
-#' - [`SimilarityMetrics()`]: Generates distance metrics from quartet statuses.
+#' - Generate distance metrics from quartet statuses: [`SimilarityMetrics()`].
 #' 
 #' @references 
 #' - \insertRef{Brodal2013}{Quartet}
@@ -58,22 +70,57 @@
 #'
 #' - \insertRef{Sand2014}{Quartet}
 #'
-#' @importFrom Rdpack reprompt 
+#' @importFrom Rdpack reprompt
+#' @importFrom ape keep.tip
+#' @importFrom TreeTools AllTipLabels TipLabels
 #' @name QuartetStatus
 #' @export
-QuartetStatus <- function (trees, cf=trees[[1]]) {
-  SingleTreeQuartetAgreement(trees, comparison = cf)
+QuartetStatus <- function (trees, cf = trees[[1]], nTip = NULL) {
+  if (is.null(nTip)) {
+    SingleTreeQuartetAgreement(trees, comparison = cf)
+  } else {
+    if (isTRUE(nTip)) nTip <- length(AllTipLabels(c(list(cf), c(trees))))
+    Q <- choose(nTip, 4)
+    status <- vapply(c(trees), function (x) {
+      commonLabels <- intersect(TipLabels(x), TipLabels(cf))
+      resolvedX <- ResolvedQuartets(x)
+      resolvedCf <- ResolvedQuartets(cf)
+      if (length(commonLabels) > 3L) {
+        reducedX <- keep.tip(x, commonLabels)
+        reducedCf <- keep.tip(cf, commonLabels)
+        resolvedReducedX <- ResolvedQuartets(reducedX)
+        resolvedReducedCf <- ResolvedQuartets(reducedCf)
+        commonStatus <- SingleTreeQuartetAgreement(reducedX, reducedCf)
+        quartetsIn1Only <- resolvedX - c(sum(commonStatus[, c('s', 'd', 'r1')]),
+                                         sum(commonStatus[, c('r2', 'u')]))
+        quartetsIn2Only <- resolvedCf - c(sum(commonStatus[, c('s', 'd', 'r2')]),
+                                          sum(commonStatus[, c('r1', 'u')]))
+        
+        commonStatus[, c('r1', 'u')] <- commonStatus[, c('r1', 'u')] + quartetsIn1Only
+        commonStatus[, c('r2', 'u')] <- commonStatus[, c('r2', 'u')] + quartetsIn2Only
+        commonStatus[, 'u'] <- commonStatus[, 'u'] + Q - sum(resolvedX, resolvedCf, -resolvedReducedX)
+        commonStatus
+      } else {
+        c(0, 0, 0, 0, resolvedX[1], resolvedCf[1], Q - resolvedX[1] - resolvedCf[1])
+      }
+    }, c('N' = 0, 'Q' = 0, 's' = 0, 'd' = 0, 'r1' = 0, 'r2' = 0, 'u' = 0))
+    status[c('N', 'Q'), ] <- c(Q + Q, Q)
+    
+    # Return:
+    t(status)
+  }
 }
 
 #' Wrapper for tqDist
 #' 
-#' Convenience function that takes a list of trees, writes them to the text file
-#' expected by the C implementation of tqDist (Sand _et al._ 2014).
-#' tqDist is then called, and the temporary file is deleted when analysis is complete.
+#' Convenience function that takes a list of trees, writes them to the text
+#' file expected by the C implementation of tqDist (Sand _et al._ 2014).
+#' tqDist is then called, and the temporary file is deleted when analysis is
+#' complete.
 #' 
-#' Quartets can be resolved in one of five ways, which 
-#'  Brodal _et al_. (2013) and Holt _et al_. (2014) distinguish using the letters
-#'  A--E, and Estabrook (1985) refers to as:
+#' Quartets can be resolved in one of five ways, which Brodal _et al_. (2013)
+#' and Holt _et al_. (2014) distinguish using the letters A--E, and
+#' Estabrook _et al._ (1985) refers to as:
 #'  
 #'  - A: _s_ = resolved the **s**ame in both trees;
 #'  
@@ -117,7 +164,7 @@ TQAE <- function (trees) {
   result <- .Call('_Quartet_tqdist_AllPairsQuartetAgreementEdge',
                   .TreeToEdge(trees))
   nTrees <- nrow(result)
-  array(result, c(nTrees, nTrees, 2), dimnames=list(NULL, NULL, c('A', 'E')))
+  array(result, c(nTrees, nTrees, 2), dimnames = list(NULL, NULL, c('A', 'E')))
 }
 
 #' Check tree size
@@ -130,45 +177,84 @@ TQAE <- function (trees) {
 #' from C.
 #' 
 #' @keywords internal
+#' @export
 .CheckSize <- function (tree) UseMethod('.CheckSize')
 
+#' @rdname dot-CheckSize
+#' @keywords internal
+#' @export
 .CheckSize.phylo <- function (tree) {
   if (length(tree$tip.label) > 477L) {
     warning("Trees with > 477 tips may produce integer overflow errors.")
   }
 }
 
+#' @rdname dot-CheckSize
+#' @export
+#' @keywords internal
 .CheckSize.list <- function (tree) {
-  .CheckSize(tree[[1]])
+  lapply(tree, .CheckSize)
 }
 
+#' @rdname dot-CheckSize
+#' @export
+#' @keywords internal
 .CheckSize.multiPhylo  <- .CheckSize.list
 
-#' @describeIn QuartetStatus Agreement of each quartet, comparing each pair of trees 
-#' in a list.
-#' @return `ManyToManyQuartetAgreement()` returns a three-dimensional array listing,
-#'   for each pair of trees in turn, the number of quartets in each category.
+#' @describeIn QuartetStatus Agreement of each quartet, comparing each pair of
+#' trees in a list.
+#' @return `ManyToManyQuartetAgreement()` returns a three-dimensional array 
+#' listing, for each pair of trees in turn, the number of quartets in each
+#' category.
 #' @examples 
-#'  # Calculate Quartet Divergence between each tree and each other tree in a 
-#'  # list
-#'  QuartetDivergence(ManyToManyQuartetAgreement(two_moved))
+#' # Calculate Quartet Divergence between each tree and each other tree in a 
+#' # list
+#' QuartetDivergence(ManyToManyQuartetAgreement(two_moved))
+#' @importFrom TreeTools PairwiseDistances
 #' @export 
-ManyToManyQuartetAgreement <- function (trees) {
+ManyToManyQuartetAgreement <- function (trees, nTip = NULL) {
   treeNames <- names(trees)
-  AE <- TQAE(trees)
-  nTree <- dim(AE)[1]
-  A   <- AE[, , 1]
-  E   <- AE[, , 2]
-  ABD <- matrix(diag(A), nTree, nTree)
-  CE  <- matrix(diag(E), nTree, nTree)
-  DE  <- t(CE)
-  C   <- CE - E
-  D   <- DE - E
-  B   <- ABD - A - D
-  
-  # Return:
-  array(c(A, B, C, D, E), dim = c(nTree, nTree, 5),
-        dimnames = list(treeNames, treeNames, c('s', 'd', 'r1', 'r2', 'u')))
+  dimNames <- list(treeNames, treeNames, c('N', 'Q', 's', 'd', 'r1', 'r2', 'u'))
+  if (is.null(nTip)) {
+    AE <- TQAE(trees)
+    nTree <- dim(AE)[1]
+    A   <- AE[, , 1]
+    E   <- AE[, , 2]
+    ABD <- matrix(diag(A), nTree, nTree)
+    CE  <- matrix(diag(E), nTree, nTree)
+    DE  <- t(CE)
+    C   <- CE - E
+    D   <- DE - E
+    B   <- ABD - A - D
+    Q   <- ABD + CE
+    N   <- Q + Q
+    
+    # Return:
+    array(c(N, Q, A, B, C, D, E), dim = c(nTree, nTree, 7),
+          dimnames = dimNames)
+  } else {
+    if (isTRUE(nTip)) nTip <- length(AllTipLabels(trees))
+    treeNames <- names(trees)
+    ret <- vapply(PairwiseDistances(trees, QuartetStatus, 7, nTip = nTip),
+                  as.matrix, matrix(0, length(trees), length(trees), 
+                                    dimnames = dimNames[1:2]))
+    dimnames(ret)[[3]] <- c('N', 'Q', 's', 'd', 'r1', 'r2', 'u')
+    
+    diag(ret[, , 'N']) <- 2L * choose(nTip, 4)
+    diag(ret[, , 'Q']) <- choose(nTip, 4)
+    
+    resolved <- vapply(trees, ResolvedQuartets, double(2))
+    diag(ret[, , 's']) <- resolved[1, ]
+    diag(ret[, , 'u']) <- choose(nTip, 4) - resolved[1, ]
+    
+    swapTri <- lower.tri(ret[, , 'r1'])
+    tmp <- ret[, , 'r1'][swapTri]
+    ret[, , 'r1'][swapTri] <- ret[, , 'r2'][swapTri]
+    ret[, , 'r2'][swapTri] <- tmp
+
+    # Return:
+    ret
+  }
 }
 
 #' @describeIn QuartetStatus Agreement of each quartet in trees in one list with
@@ -178,9 +264,9 @@ ManyToManyQuartetAgreement <- function (trees) {
 #' @return `TwoListQuartetAgreement()` returns a three-dimensional array listing,
 #'   for each pair of trees in turn, the number of quartets in each category.
 #' @examples 
-#'   # Calculate Quartet Divergence between each tree in one list and each 
-#'   # tree in another
-#'   QuartetDivergence(TwoListQuartetAgreement(sq_trees[1:3], sq_trees[10:13]))
+#' # Calculate Quartet Divergence between each tree in one list and each 
+#' # tree in another
+#' QuartetDivergence(TwoListQuartetAgreement(sq_trees[1:3], sq_trees[10:13]))
 #' @export
 TwoListQuartetAgreement <- function (trees1, trees2) {
   aperm(vapply(trees2, function (cf) SingleTreeQuartetAgreement(trees1, cf),
@@ -188,8 +274,8 @@ TwoListQuartetAgreement <- function (trees1, trees2) {
   
 }
 
-#' @describeIn QuartetStatus Agreement of each quartet in trees in a list with the
-#' quartets in a comparison tree.
+#' @describeIn QuartetStatus Agreement of each quartet in trees in a list with
+#' the quartets in a comparison tree.
 #' @param comparison A tree of class \code{\link[ape:read.tree]{phylo}} against
 #' which to compare `trees`.
 #' @return `SingleTreeQuartetAgreement()` returns a two-dimensional array listing,
@@ -226,7 +312,7 @@ SingleTreeQuartetAgreement <- function (trees, comparison) {
   
   # Return:
   array(c(rep(2L * Q, nTree), rep(Q, nTree), A, B, C, D, E), dim=c(nTree, 7L),
-        dimnames=list(names(trees), c('N', 'Q', 's', 'd', 'r1', 'r2', 'u')))
+        dimnames = list(names(trees), c('N', 'Q', 's', 'd', 'r1', 'r2', 'u')))
 }
 
 #' tqDist file generator
@@ -254,9 +340,10 @@ TQFile <- function (treeList) {
   fileName
 }
 
-#' Triplet and quartet distances with tqDist
+#' Direct entry points to 'tqDist' functions
 #' 
-#' Functions to calculate triplet and quartet distances between pairs of trees.
+#' Wrappers for functions in 'tqDist', which calculate triplet and quartet
+#' distances between pairs of trees.
 #' 
 #' @param file,file1,file2 Paths to files containing a tree or trees in Newick
 #'  format, possibly created using [`TQFile()`].
@@ -374,16 +461,19 @@ AllPairsQuartetDistance <- function(file) {
   }
 }
 
-#' @importFrom TreeTools RenumberTips RenumberTree
 #' @keywords internal
 #' @export
 .TreeToEdge <- function (trees, tipOrder) UseMethod('.TreeToEdge')
 
+#' @keywords internal
 .TreeToEdge.list <- function (trees, tipOrder = trees[[1]]$tip.label) {
   lapply(trees, .SortTree, tipOrder)
 }
+#' @keywords internal
 .TreeToEdge.multiPhylo <- .TreeToEdge.list
 
+#' @keywords internal
+#' @importFrom TreeTools RenumberTips RenumberTree
 .TreeToEdge.phylo <- function (trees, tipOrder = NULL) {
   if (is.null(tipOrder)) {
     edge <- trees$edge