+++ src/rQuartetDist.cpp
@@ -6,7 +6,7 @@ using namespace Rcpp;
 
 #include <vector>
 
-//' Triplet and quartet distances with tqDist
+//' Direct entry points to 'tqDist' functions
 //' 
 //' Functions to calculate triplet and quartet distances between pairs of trees.
 //' Input is not checked for sanity.