+++ tests/testthat/test-QuartetDistance.R
@@ -15,7 +15,7 @@ test_that("Distances are calculated from strings", {
   expect_equal(AllPairsQuartetAgreement(fileName),
                TQAE(trees))
   expect_equal(ManyToManyQuartetAgreement(trees)[1, , ],
-               SingleTreeQuartetAgreement(trees, trees[[1]])[, 3:7])
+               SingleTreeQuartetAgreement(trees, trees[[1]])[, ])
 })
 
 test_that("Splits are compared", {